delete line honors selected text. ^H replace on correct sub
The ^L delete line shortcut now deletes all lines with selected text. If the cursor is on the start of a line and no text is selected on that line, that line is also deleted.

I had previously added the ^H shortcut to the wrong code block in ide_methods.bas. I moved that to the correct code block (the same as Alt-F3) so ^H now opens the Search And Replace dialog, consistent with other code editors on Windows.